Extract out code for naming styles into its own class.
