+++ src/Workspaces/Core/Portable/Workspaces.csproj
@@ -390,6 +390,7 @@
     <Compile Include="NamingStyles\Serialization\NamingStylePreferencesExtensions.cs" />
     <Compile Include="Options\EditorConfig\EditorConfigStorageLocation.cs" />
     <Compile Include="Options\EditorConfig\IEditorConfigStorageLocation.cs" />
+    <Compile Include="Options\EditorConfig\NamingStylePreferenceEditorConfigStorageLocation.cs" />
     <Compile Include="PatternMatching\PatternMatch.cs" />
     <Compile Include="PatternMatching\PatternMatcher.AllLowerCamelCaseMatcher.cs" />
     <Compile Include="PatternMatching\PatternMatcher.cs" />
