+++ src/Workspaces/CoreTest/EditorConfigStorageLocation/EditorConfigStorageLocationTests.cs
@@ -16,7 +16,7 @@ namespace Microsoft.CodeAnalysis.UnitTests.EditorConfig.StorageLocation
         [Fact]
         public static void TestEmptyDictionaryReturnNoNamingStylePreferencesObjectReturnsFalse()
         {
-            var editorConfigStorageLocation = new EditorConfigStorageLocation<NamingStylePreferences>();
+            var editorConfigStorageLocation = new NamingStylePreferenceEditorConfigStorageLocation();
             var result = editorConfigStorageLocation.TryGetOption(new object(), new Dictionary<string, object>(), typeof(NamingStylePreferences), out var @object);
             Assert.False(result, "Expected TryParseReadonlyDictionary to return 'false' for empty dictionary");
         }
@@ -24,7 +24,7 @@ namespace Microsoft.CodeAnalysis.UnitTests.EditorConfig.StorageLocation
         [Fact]
         public static void TestEmptyDictionaryDefaultNamingStylePreferencesObjectReturnsFalse()
         {
-            var editorConfigStorageLocation = new EditorConfigStorageLocation<NamingStylePreferences>();
+            var editorConfigStorageLocation = new NamingStylePreferenceEditorConfigStorageLocation();
             var existingNamingStylePreferences = new NamingStylePreferences(
                 ImmutableArray.Create<SymbolSpecification>(),
                 ImmutableArray.Create<NamingStyle>(),
@@ -53,7 +53,7 @@ namespace Microsoft.CodeAnalysis.UnitTests.EditorConfig.StorageLocation
             };
             var existingNamingStylePreferences = ParseDictionary(initialDictionary);
 
-            var editorConfigStorageLocation = new EditorConfigStorageLocation<NamingStylePreferences>();
+            var editorConfigStorageLocation = new NamingStylePreferenceEditorConfigStorageLocation();
             var newDictionary = new Dictionary<string, object>()
             {
                 ["dotnet_naming_rule.methods_and_properties_must_be_pascal_case.severity"] = "error",
@@ -79,7 +79,7 @@ namespace Microsoft.CodeAnalysis.UnitTests.EditorConfig.StorageLocation
         [Fact]
         public static void TestObjectTypeThrowsNotSupportedException()
         {
-            var editorConfigStorageLocation = new EditorConfigStorageLocation<NamingStylePreferences>();
+            var editorConfigStorageLocation = new NamingStylePreferenceEditorConfigStorageLocation();
             Assert.Throws<NotSupportedException>(() =>
             {
                 editorConfigStorageLocation.TryGetOption(new object(), new Dictionary<string, object>(), typeof(object), out var @object);
